Clean up remote Instruments properly
Shared instruments can (obviously) be shared, so we need to keep track
of their refcount. We should only try to close them if the client which
is disconnecting is the last remaining client. We also need to remove
them from the shared object table so they'll be re-created when opened
by another client.