Change how Instruments are initialized
This removes some of the crazier magic from the metaclass and makes it a bit more explicit. Now
driver developers are forbidden from defining an __init__ method, and instead must implement a
method called `_initialize`. This method is called with _paramset (and other mixin-related
attributes) already set, and is passed in any settings as keyword arguments.