Add a way to add special resource-closing code for VISA resources
Now if you add a `_close_resource` static/classmethod to an instrument
class, `list_instruments` will invoke it when closing the resource. This
is useful if you need to run some extra code to e.g. disable 'remote'
mode on the device. The ordinary `close` method is unsuitable here
because list_instruments doesn't actually instantiate an Instrument.