Skip listing visa instruments if there's no matching module
Originally, we found all visa instruments, then filtered out the ones
that didn't match the 'module' argument. Many times this lookup is
unnecessary, so this change first checks if there are *any* visa drivers
that match the given module string (and aren't in the blacklist). If
not, we skip list_visa_instruments altogether.