Don't mock numpy.
This isn't necessary (since ReadTheDocs should be installing numpy anyway) and was causing issues
with pint's compat checks. Pint was importing pandas (which I had installed), which imported numpy
and checked its version string. Since numpy was mocked, its __version__ was a Mock rather than a
string, which was causing errors. This was probably the cleanest way to fix the issue.